refactor(wire): extract AttachmentBuilder internal helper
